# docs/05-openapi.yaml
openapi: 3.1.0
info:
  title: Kiksu API — Phase 1 spine
  version: "1.0.0"
  summary: Onboarding, verification, session, timetable, forum and reviews for Kiksu.
  description: |
    Server-layer HTTP contract for Kiksu, the anonymous app for university students in
    Azerbaijan.

    **Read `docs/05-api-conventions.md` first.** It is binding on every operation here:
    versioning and app-store rollout (§1), the JWT claim allowlist and the `user_metadata`
    trap (§2), the closed error code set and the deliberately uniform codes (§3), keyset
    cursors (§4), rate limits (§5), idempotency (§6), and — most importantly — §8, the
    field-by-field prohibition on any response that links two identity layers.

    **Scope.** Phase 1 only: onboarding and verification (all three routes), session,
    profile and privacy, timetable (catalogue, sections, week grid, attendance, deadlines,
    GPA), forum (boards, feeds, posts, comments, votes, polls, alias preview), and reviews
    (professor and course profiles, review list, write review, contribution wall).
    Marketplace, vacancies, careers, events and clubs are Phase 2/3 and are deliberately
    absent. There are **zero** career-layer fields anywhere in this document, by
    `02-identity-spec.md` assertion 3.

    **Servability.** Every response field traces to a column, a view or a documented server
    computation in `docs/01-schema.sql`. Fields that are computed rather than stored say so
    in their description.

...
  contact:
...
servers:
...
security:
  - supabaseJwt: []
...
paths:
...
  /catalogue/courses:
    get:
      tags: [catalogue]
      operationId: searchCourses
      summary: Course catalogue search
      description: |
        Full-text plus trigram over the folded title and code. Folding maps
        `ə→e ğ→g ı→i ö→o ş→s ü→u ç→c`, so `"verilenler"` matches
        `"Verilənlər bazası sistemləri"` and `"Eliyeva"` matches `"Əliyeva"`. The query is
        folded identically to the index — folding one side only is the classic way to break
        Azerbaijani search.

        Scoped to the caller's university.
      parameters:
        - name: q
          in: query
          schema: { type: string, minLength: 2, examples: ["verilenler"] }
        - name: department_id
          in: query
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Courses.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CoursePage" }
              examples:
                cs214:
                  value:
                    items:
                      - id: "0193f350-a0b1-7c2d-e3f4-5a6b7c8d9e0f"
                        code: "CS 214"
                        title: "Verilənlər bazası sistemləri"
                        short_title: "Verilənlər bazası"
                        credits: 6
                        level: 2
                        department_name: "İnformatika kafedrası"
                        review_summary:
                          review_count: 61
                          rating_avg: 4.2
                    page_info: { next_cursor: null, has_more: false }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /catalogue/instructors:
    get:
      tags: [catalogue]
      operationId: searchInstructors
      summary: Instructor search
      description: |
        Instructor names are **real and public** — this is a professor-review product.
        Instructors are not `app_user`s and are never linked to one. This endpoint has no
        relationship to handle search and no relationship to the k-anonymity floor.
      parameters:
        - name: q
          in: query
          schema: { type: string, minLength: 2, examples: ["Eliyeva"] }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Instructors.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstructorPage" }
        "429": { $ref: "#/components/responses/RateLimited" }

  # ==================================================================
  # TIMETABLE
  # ==================================================================
...
  /posts:
    post:
      tags: [forum]
      operationId: createPost
      summary: Create a thread
      description: |
        **The author always receives alias ordinal 1** — the counter starts at 0 and
        allocation runs for the author first, inside the same transaction as the content
        insert (`02-identity-spec.md` §3.4, assertion 15). The client does not send an
        ordinal and cannot influence it.

        `attachment_upload_ids` reference objects already uploaded through a signed target
        and already **re-encoded server-side with all metadata destroyed** — a re-encode,
        not a tag strip, because thumbnails embedded in EXIF survive naive stripping
        (T7, assertion 29). A post cannot reference an attachment that has not finished
        processing.

        Rate limited by `forum.post.create`, which scales by tier *and* account age: a
        provisional account gets 3/day, and an account under 24 hours old gets a quarter of
        its tier's allowance.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [board_id, title]
              properties:
                board_id: { type: string, format: uuid }
                kind: { $ref: "#/components/schemas/PostKind" }
                title: { type: string, minLength: 1, maxLength: 200 }
                body: { type: string, maxLength: 20000 }
                lang: { $ref: "#/components/schemas/LocaleCode" }
                attachment_upload_ids:
                  type: array
                  maxItems: 4
                  items: { type: string, format: uuid }
                poll: { $ref: "#/components/schemas/PollInput" }
            examples:
              text:
                value:
                  board_id: "0193f3b0-c8d9-7e05-a2b3-4c5d6e7f8091"
                  kind: "text"
                  title: "Aralıq imtahanı təxirə salındı, kim eşitdi?"
                  body: "Dekanlıqdan elan gəldi deyirlər, amma qrupda heç kim dəqiq bilmir."
                  lang: "az"
              poll:
                value:
                  board_id: "0193f3b0-c8d9-7e05-a2b3-4c5d6e7f8091"
                  kind: "poll"
                  title: "CS 214 üçün hansı gün konsultasiya olsun?"
                  lang: "az"
                  poll:
                    question: "Hansı gün?"
                    is_multi_choice: false
                    max_choices: 1
                    closes_at: "2026-08-17T20:00:00+04:00"
                    options: ["Çərşənbə axşamı", "Cümə axşamı", "Şənbə"]
      responses:
        "201":
          description: Created. `author.alias_number` is always 1.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostDetail" }
        "403":
          description: "`verification_required`, `tier_insufficient`, `sanction_active` or `board_archived`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429":
          description: "`daily_post_quota` or `rate_limited`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                quota:
                  value:
                    error:
                      code: "daily_post_quota"
                      message: "Bu gün üçün post limitinə çatdın. Sabah yenidən yaza bilərsən."
                      message_key: "err.daily_post_quota"
                      action: "wait_cooldown"
                      action_label: null
                      details: {}
                      request_id: "01JAV2H5X3VR0Y2CH4B28S9G1J"
...
  /search/posts:
    get:
      tags: [search]
      operationId: searchPosts
      summary: Forum post search (HM-04)
      description: |
        Full-text over `post.search_vector`, campus-scoped and gated on the board's
        `min_tier_to_read` exactly as the board feed is. A card-gated board's contents
        are not reachable here by an email-tier caller — otherwise search would read
        what the board list refuses even to name.

        **Language.** Each row's vector was built with that row's own `lang` config
        (`util.az` is `simple`, since no Azerbaijani stemmer exists; `util.ru` and
        `util.en` carry real stemmers). The query is evaluated as three constant arms
        paired to `lang`, so all three languages match and the GIN index still serves
        the predicate.

        **Folding.** `ə→e ğ→g ı→i ö→o ş→s ü→u ç→c`, plus the combining dot above that
        `lower('İ')` produces. Applied to the query through `util.tsq()` exactly as it
        was to the index. Folding one side only does not degrade results, it returns
        none.

        Results carry an author alias and a coarse tier, never a handle (P12).
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 120, examples: ["imtahan"] }
        - name: scope
          in: query
          schema: { type: string, enum: [all, campus, national], default: all }
          description: Separates the national layer from the caller's campus boards.
        - name: board
          in: query
          schema: { type: string, maxLength: 80 }
          description: Board slug. Narrows to one board; there is no author filter (P12).
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Matching posts, most relevant first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PostHitPage" }
        "400": { $ref: "#/components/responses/CursorInvalid" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /search/courses:
    get:
      tags: [search]
      operationId: searchCourseCatalogue
      summary: Course search with rating summary (HM-05)
      description: |
        Three match paths OR'd together: the full-text vector, a folded code prefix so
        `cs2` reaches `CS 214`, and a folded substring over the title. A code-prefix hit
        outranks a body mention — someone typing `CS 214` wants that course, not every
        course that mentions it.

        Scoped to the caller's university. The rating summary is inline so the result
        list is useful without a tap.

        **On the two course-search paths.** `/catalogue/courses`, described further up
        this document, was never implemented at that path; the catalogue search that
        exists is `GET /v1/timetable/courses`, which returns a different projection
        (instructor names, no ratings) for the timetable's course picker. This endpoint
        is the canonical one for search and returns ratings instead.

        They are two queries over one corpus, which is a drift risk worth naming: both
        apply the same campus scope and the same three fold predicates, so a change to
        the fold rules has to be made in both `SearchService.searchCourses` and
        `TimetableService.searchCourses` or Azerbaijani search quietly starts behaving
        differently depending on which screen the student came from.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 120, examples: ["verilenler"] }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Matching courses.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CourseHitPage" }
              examples:
                cs214:
                  value:
                    items:
                      - id: "0193f350-a0b1-4c2d-e3f4-5a6b7c8d9e0f"
                        code: "CS 214"
                        title: "Verilənlər bazası sistemləri"
                        credits: 6
                        department: "İnformatika"
                        rating_avg: 4.34
                        review_count: 61
                    next_cursor: null
        "400": { $ref: "#/components/responses/CursorInvalid" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /search/instructors:
    get:
      tags: [search]
      operationId: searchInstructorDirectory
      summary: Instructor search with rating summary (HM-05)
      description: |
        Instructor names are **real and public** — this is a professor-review product,
        and an instructor is never an `app_user`. This endpoint has no relationship to
        handle search and none to the k-anonymity floor.

        The trigram index over `name_folded` is what makes `Eliyeva` find `Əliyeva`.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 120, examples: ["Eliyeva"] }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Matching instructors.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InstructorHitPage" }
        "400": { $ref: "#/components/responses/CursorInvalid" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /search/listings:
    get:
      tags: [search]
      operationId: searchListings
      summary: Marketplace search (HM-06)
      description: |
        Campus-scoped like the browse feed. Carries the seller's handle: the marketplace
        is Layer 2 always, because a seller with no persistent reputation is a scam
        waiting to happen. That is also why this cannot share a response body with
        `/search/posts` — see the section note above.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 120 }
        - name: sort
          in: query
          schema: { type: string, enum: [relevance, price_asc, price_desc, newest], default: relevance }
          description: |
            Bound into the cursor. Changing `sort` while presenting a cursor is
            `400 cursor_invalid`, not a silently re-keyed page.
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Matching listings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListingHitPage" }
        "400": { $ref: "#/components/responses/CursorInvalid" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /search/vacancies:
    get:
      tags: [search]
      operationId: searchVacancies
      summary: Vacancy search (HM-06)
      description: |
        National unless the employer targeted specific campuses through
        `target_university_ids`. Vacancies with no `external_url` are excluded: Kiksu
        hands off rather than taking applications, so a vacancy with no link is one
        nobody can act on.

        `sort=deadline` orders soonest-first and keeps undated vacancies at the end
        rather than dropping them.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 120 }
        - name: sort
          in: query
          schema: { type: string, enum: [relevance, deadline, newest], default: relevance }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Matching vacancies.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VacancyHitPage" }
        "400": { $ref: "#/components/responses/CursorInvalid" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
components:
  securitySchemes:
    supabaseJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase-issued access token, **ES256**, 900 s TTL, refresh-token rotation with reuse
        detection.

        The algorithm was verified against a real token from the live project — the header
        says ES256 and the JWKS publishes a single EC key. `JwtVerifierService` deliberately
        pins no algorithm and resolves the key by `kid` from the published set, so a key
        rotation or curve change on Supabase's side is a non-event rather than every token
        being rejected at once.

        **Trusted claims** (`05-api-conventions.md` §2.2): `sub`, and from `app_metadata`
        only — `app_user_id`, `tier`, `role`, `univ_id`, `epoch`, `sid`.

        **Never trusted:** `user_metadata` / `raw_user_meta_data` in any form. It is
        **client-writable** through the standard Supabase update-user call and it lands in
        the JWT correctly signed. A policy that trusts `user_metadata.tier` grants every user
        the ability to award themselves the `ANONİM KART` badge. Also never trusted: `email`,
        `phone`, or any claim naming faculty, entry year, study year, handle, karma or a
        career reference.

        Never present the token in a URL.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: |
        Client-generated UUIDv4, one per logical user intent — generated when the composer
        opens, not when the request fires, so a retry after a network failure reuses it.
        Scoped to `(app_user_id, method, path, key)`, retained 24 h. A replay returns the
        stored response with `Idempotency-Replayed: true`.
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      description: Clamped, never rejected.
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: |
        **Opaque.** Present it back exactly as received. It is an HMAC-signed keyset tuple
        bound to the endpoint and to every sort/filter parameter; a modified or hand-crafted
        cursor returns `400 cursor_invalid`, which is what stops cursor-crafting being used
        as a scan primitive against forum publication times. Passing `cursor` together with a
        changed `tab`/`sort`/`q`/filter is also `cursor_invalid`. Cursors expire after
        15 minutes. Never parse it.
...
  headers:
    RateLimitLimit:
      description: Requests permitted in the current window for the scarcest applicable bucket.
      schema: { type: integer }
    RateLimitRemaining:
      description: Requests remaining in that bucket.
      schema: { type: integer }
    RateLimitReset:
      description: Seconds until that bucket resets. A delta, never an absolute time.
      schema: { type: integer }
    RateLimitPolicy:
      description: Named policy, e.g. `"forum-write";q=15;w=86400`. Names the bucket, never the principal.
      schema: { type: string }
    RetryAfter:
      description: Seconds to wait. Clients must honour it and must not busy-retry.
      schema: { type: integer }
...
  responses:
    CursorInvalid:
      description: |
        `cursor_invalid`. The cursor was tampered with, expired, or was presented
        alongside a changed sort or filter. Deliberately undifferentiated between those
        causes (§3.4): a distinguishable failure reason would help an attacker iterate
        toward a forged cursor, and a forged cursor is a scan primitive against forum
        publication times (T9). The client discards it and reloads from the first page.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            invalid:
              value:
                error:
                  code: "cursor_invalid"
                  message: "Səhifələmə etibarsızdır. Siyahı yenidən yüklənir."
                  message_key: "err.cursor_invalid"
                  action: "retry"
                  action_label: "Yenidən cəhd et"
                  details: {}
                  request_id: "01JAV2H7Z5XT2A4EK6D4AU1J3M"
    Unauthenticated:
      description: "`unauthenticated`, `token_stale` or `token_invalid`. On `token_stale` the client refreshes and retries once, silently."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            stale:
              value:
                error:
                  code: "token_stale"
                  message: "Sessiya yeniləndi. Yenidən cəhd edirik."
                  message_key: "err.token_stale"
                  action: "refresh_token"
                  action_label: null
                  details: {}
                  request_id: "01JAV2H7Z5XT2A4EK6D4AU1J3M"
...
    ValidationFailed:
      description: "`validation_failed`, with `details.fields`."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            range:
              value:
                error:
                  code: "validation_failed"
                  message: "Qiymətləndirmə 1 ilə 5 arasında olmalıdır."
                  message_key: "err.validation_failed"
                  action: "none"
                  action_label: null
                  details:
                    fields:
                      - { path: "overall_rating", rule: "range", min: 1, max: 5 }
                  request_id: "01JAV2H917ZV4C6GM8F6CW3L5P"
    RateLimited:
      description: "`rate_limited` and its domain-specific siblings. Honour `Retry-After`."
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
        RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
        RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
        RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
        RateLimit-Policy: { $ref: "#/components/headers/RateLimitPolicy" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            slowDown:
              value:
                error:
                  code: "rate_limited"
                  message: "Çox sürətli göndərirsən. 3 dəqiqədən sonra yenidən cəhd et."
                  message_key: "err.rate_limited"
                  action: "retry_after"
                  action_label: null
                  details: {}
                  request_id: "01JAV2HA280W5D7HN9G7DX4M6Q"
...
  schemas:
...
    PostHitPage:
      type: object
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/PostHit" } }
        next_cursor: { type: string, nullable: true }
...
    CourseHitPage:
      type: object
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/CourseHit" } }
        next_cursor: { type: string, nullable: true }
...
    InstructorHitPage:
      type: object
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/InstructorHit" } }
        next_cursor: { type: string, nullable: true }
...
    ListingHitPage:
      type: object
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ListingHit" } }
        next_cursor: { type: string, nullable: true }
...
    VacancyHitPage:
      type: object
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/VacancyHit" } }
        next_cursor: { type: string, nullable: true }

    # ---------- primitives ----------
    LocaleCode:
      type: string
      enum: [az, ru, en]
      description: Default `az`.
...
    PostKind:
      type: string
      enum: [text, image, link, poll]
      x-kiksu-enum: open
...
    PageInfo:
      type: object
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: [string, "null"]
          description: Opaque. `null` exactly when `has_more` is false.
        has_more:
          type: boolean
          description: |
            The loop condition. Never infer the end from an empty `items` array — a page can
            legitimately come back short after moderation filtering.
      description: No `total_count` exists anywhere in this API. A total is an expensive scan, and on people-shaped collections it is a cohort count.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, message_key, request_id]
          properties:
            code:
              type: string
              description: From the closed set in `05-api-conventions.md` §3.3. Safe to `switch` on.
            message:
              type: string
              description: Localised, student-facing, actionable. Never a stack trace, a constraint name, an internal service name or an identifier.
            message_key:
              type: string
              description: Stable key into the client's own catalogue. The client prefers its own string when it knows the key, which is what makes adding a code non-breaking for old builds.
            action:
              type: [string, "null"]
              enum: [none, retry, retry_after, sign_in, refresh_token, verify_email, verify_card, upgrade_app, open_settings, wait_cooldown, contact_support, null]
            action_label: { type: [string, "null"] }
            details:
              type: object
              additionalProperties: true
              description: |
                Code-specific. **Never contains a count of people** — not a cohort size, not a
                follower count below 1,000, not a "you are one of N" hint
                (`02-identity-spec.md` assertion 35).
            request_id: { type: string }
            store_url:
              type: [string, "null"]
              description: Present only on `client_version_unsupported`, whose shape is frozen forever.

    # ---------- meta ----------
...
    Course:
      type: object
      required: [id, code, title]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, examples: ["CS 214"] }
        title: { type: string, examples: ["Verilənlər bazası sistemləri"] }
        short_title: { type: [string, "null"], examples: ["Verilənlər bazası"], description: The grid label. }
        credits: { type: [number, "null"], examples: [6] }
        level: { type: [integer, "null"] }
        description: { type: [string, "null"] }
        department_name: { type: [string, "null"], examples: ["İnformatika kafedrası"] }
        review_summary:
          type: [object, "null"]
          description: From `public.course_review_summary`. Never walled.
          properties:
            review_count: { type: integer }
            rating_avg: { type: [number, "null"] }
    CoursePage:
      type: object
      required: [items, page_info]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Course" }
        page_info: { $ref: "#/components/schemas/PageInfo" }
    Instructor:
      type: object
      required: [id, full_name]
      properties:
        id: { type: string, format: uuid }
        full_name:
          type: string
          examples: ["Nigar Əliyeva"]
          description: Real and public. Instructors are not `app_user`s and are never linked to one.
        title_prefix: { type: [string, "null"], examples: ["dos."] }
        initials: { type: [string, "null"], examples: ["NƏ"], description: Rendered avatar. }
        department_name: { type: [string, "null"], examples: ["İnformatika kafedrası"] }
        university_code: { type: [string, "null"], examples: ["BDU"] }
        is_claimed: { type: boolean }
        rating_avg: { type: [number, "null"] }
        review_count: { type: [integer, "null"] }
    InstructorPage:
      type: object
      required: [items, page_info]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Instructor" }
        page_info: { $ref: "#/components/schemas/PageInfo" }
...
    PollInput:
      type: object
      required: [options]
      properties:
        question: { type: string, maxLength: 300 }
        is_multi_choice: { type: boolean, default: false }
        max_choices: { type: integer, minimum: 1, default: 1 }
        closes_at: { type: string, format: date-time }
        hide_results_until_vote: { type: boolean, default: false }
        options:
          type: array
          minItems: 2
          maxItems: 8
          items: { type: string, minLength: 1, maxLength: 120 }
    PostDetail:
      allOf:
        - $ref: "#/components/schemas/PostListItem"
        - type: object
          properties:
            body: { type: [string, "null"] }
            lang: { $ref: "#/components/schemas/LocaleCode" }
            edited: { type: boolean, description: Boolean only. A raw `edited_at` would reintroduce a precise activity timestamp. }
            is_mine: { type: boolean, description: Resolved server-side from `internal.post_author`, scoped to the token. }
            attachments:
              type: array
              items: { $ref: "#/components/schemas/Attachment" }
            poll:
              oneOf:
                - $ref: "#/components/schemas/Poll"
                - type: "null"
...